+++ test/integration/runtime.js
@@ -3,6 +3,8 @@
 const { fork } = require('child_process')
 const path = require('path')
 
+const WebSocket = require('ws')
+
 const { delay } = require('./utils')
 
 const DEFAULT_TIMEOUT = 5000
@@ -25,8 +27,7 @@ const makeSuite = (description, fn) => {
 	const cases = []
 
 	const test = (description, fn) => {
-		const run = () => runWithTimeout(fn, DEFAULT_TIMEOUT)
-			.then(fn)
+		const run = config => runWithTimeout(() => fn(config), DEFAULT_TIMEOUT)
 			.then(() => {
 				console.log(`\n\x1b[32m * OK   - ${description}\x1b[0m`)
 			})
@@ -38,12 +39,12 @@ const makeSuite = (description, fn) => {
 		cases.push(run)
 	}
 
-	const runSerial = () => {
+	const runSerial = (config) => {
 		console.log(description)
 		return cases.reduce(
 			(prev, t) => prev
-				.then(t)
-				.catch(t),
+				.then(() => t(config))
+				.catch(() => t(config)),
 			Promise.resolve()
 		)
 	}
@@ -57,7 +58,9 @@ const makeSuite = (description, fn) => {
 			console.log('Error', error)
 			server.kill()
 		})
-		delay(500)()
+		const open = openConnection(PORT)
+		const run = runConnection(open)
+		delay(500) ({ open, run })
 			.then(runSerial)
 			.then(() => {
 				server.kill()
@@ -73,3 +76,36 @@ const makeSuite = (description, fn) => {
 }
 
 module.exports = makeSuite
+
+
+const openConnection = port => () => new Promise((resolve, reject) => {
+	const ws = new WebSocket(`ws://localhost:${port}`)
+	ws.onopen = () => {
+		resolve(ws)
+	}
+	ws.onerror = error => {
+		ws.close()
+		reject(error)
+	}
+})
+
+const runConnection = openConnection => (messages, timeout = 500) => {
+	const messagesList = Array.isArray(messages)
+		? messages
+		: [messages]
+	return openConnection()
+		.then(ws => new Promise((resolve, reject) => {
+			const received = []
+			setTimeout(() => {
+				ws.close()
+				resolve(received)
+			}, timeout)
+			ws.onmessage = event => {
+				received.push(JSON.parse(event.data))
+			}
+			ws.onerror = reject
+			messagesList.forEach(m => {
+				ws.send(m)
+			})
+		}))
+}
